+++ NEWS.md
@@ -79,6 +79,9 @@ Library improvements
 
   * Added `unique!` which is an inplace version of `unique` ([#20549]).
 
+  * `@test isequal(x, y)` and `@test isapprox(x, y)` now prints an evaluated expression when
+    the test fails ([#22296]).
+
 Compiler/Runtime improvements
 -----------------------------